+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/ri/web/drivers/Sdmx21Driver2.java
@@ -78,6 +78,16 @@ public final class Sdmx21Driver2 implements SdmxWebDriver {
                     .build())
             .source(SdmxWebSource
                     .builder()
+                    .name("ILO")
+                    .description("International Labour Office")
+                    .driver(RI_SDMX_21)
+                    .endpointOf("https://www.ilo.org/sdmx/rest")
+                    .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
+                    .websiteOf("https://ilostat.ilo.org/data/")
+                    .monitorOf("UptimeRobot", "m783847083-609d3e4ebc1da9455baeb63e")
+                    .build())
+            .source(SdmxWebSource
+                    .builder()
                     .name("IMF_SDMX_CENTRAL")
                     .description("International Monetary Fund SDMX Centra")
                     .driver(RI_SDMX_21)